#410 switch to gdal 3

+++ nansat/nansat.py
@@ -1560,7 +1560,7 @@ class Nansat(Domain, Exporter):
         for pn in range(len(pix[1:])):
             px0, px1 = pix[pn], pix[pn+1]
             py0, py1 = lin[pn], lin[pn+1]
-            length = np.round(np.hypot(px1-px0, py0-py1))
+            length = int(np.round(np.hypot(px1-px0, py0-py1)))
             pixVector += list(np.linspace(px0, px1, length+1)[1:])
             linVector += list(np.linspace(py0, py1, length+1)[1:])
